--- data/run.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-./compile.sh
-./main
