+++ markov_lib/core/MarkovAlgorithm.cpp
@@ -1,8 +1,11 @@
 #include "MarkovAlgorithm.h"
 
-MarkovAlgorithm::MarkovAlgorithm(std::string alphabet, std::string tuple, std::vector<MarkovCommand> commands)
-    : alphabet(std::move(alphabet))
-    , tuple(std::move(tuple))
-    , commands(std::move(commands))
+namespace markov_lib
 {
+    MarkovAlgorithm::MarkovAlgorithm(std::string alphabet, std::string tuple, std::vector<MarkovCommand> commands)
+        : alphabet(std::move(alphabet))
+        , tuple(std::move(tuple))
+        , commands(std::move(commands))
+    {
+    }
 }
