+++ markov_lib/core/MarkovAlgorithm.h
@@ -3,21 +3,24 @@
 #include <string>
 #include <vector>
 
-struct MarkovCommand {
-    std::string first;
-    std::string second;
-    int num;
-    bool is_end;
-};
+namespace markov_lib
+{
+    struct MarkovCommand {
+        std::string first;
+        std::string second;
+        int num;
+        bool is_end;
+    };
 
-class MarkovAlgorithm {
-protected:
-    std::string alphabet, tuple;
-    std::vector<MarkovCommand> commands;
+    class MarkovAlgorithm {
+    protected:
+        std::string alphabet, tuple;
+        std::vector<MarkovCommand> commands;
 
-public:
-    MarkovAlgorithm(std::string alphabet, std::string tuple, std::vector<MarkovCommand> commands);
+    public:
+        MarkovAlgorithm(std::string alphabet, std::string tuple, std::vector<MarkovCommand> commands);
 
-    virtual std::string execute(std::string input) = 0;
-    virtual std::string data() const = 0;
-};
+        virtual std::string execute(std::string input) = 0;
+        virtual std::string data() const = 0;
+    };
+}
